+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## [0.2.1] - TBD
+
+- Update `solvis` schema to latest version 0.9.2
+- Update `toshi` schema to latest version 0.3.2
+- Update `kororaa` schema to latest version 0.10.0
+- Update `hazard` schema to latest version 0.3.0
+- Upgrade serverless to v4
+- Upgrade to yarn2
+- Remove serverless-domain-manager
+
 ## [0.2.0] - 2025-07
 
 ### Added